*** empty log message ***
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk/M2@1603 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ Macaulay2/dumpdata/dumpdata.c
@@ -81,7 +81,7 @@ static void trim(char *s) {
 }
 
 static int extend_memory(void *newbreak) {
-  if (ERROR == brk(newbreak)) {
+  if (ERROR == (int)brk(newbreak)) {
     warning("loaddata: out of memory (extending break to 0x%p)", newbreak);
     return ERROR;
   }
